Install redux-logger, no default state in store
Redux logger helps understand what is going on inside your store by logging each action, the state before it and after. Also if action results in error you have it there.

Setting default is not a best idea, as you have to keep it's details in sync with the implementation of reducers (e.g. the id field must match). It's better to use actions to inject initial data into the app. I'll demonstrate it in the next commit.