Use ADD_TASK action to fill state with initial set of tasks
The description in data/tasks.js is stripped from all properties, except the ones expected by action creator (taskType and text).

In todoapp.js, before the Provider and Router are initialized, actions are dispatch to fill the state with tasks. I leave the implementation of habits and progressbars to you.